[add] CPU側の実装を追加
