+++ buffer/base/build.gradle.kts
@@ -3,11 +3,18 @@ import org.gradle.nativeplatform.platform.internal.DefaultNativePlatform
 plugins {
     kotlin("multiplatform")
     alias(libs.plugins.serialization)
+    alias(libs.plugins.android.kmp.library)
 }
 
 kotlin {
     applyDefaultHierarchyTemplate()
     jvm()
+
+    android {
+        compileSdk = 35
+        minSdk = 28
+        namespace = "com.wsr.base"
+    }
 
     val hostOs = DefaultNativePlatform.getCurrentOperatingSystem()
     val hostArch = DefaultNativePlatform.getCurrentArchitecture()
