+++ buffer/cpu/src/androidMain/kotlin/com/wsr/cpu/CPUBackend.android.kt
@@ -1,0 +1,8 @@
+package com.wsr.cpu
+
+import com.wsr.base.IBackend
+
+actual fun loadCPUBackend(): IBackend? = runCatching<IBackend> {
+    System.loadLibrary("cpu")
+    CPUJvmBackend()
+}.getOrNull()
