+++ buffer/cpu/src/jvmAndroidMain/kotlin/com/wsr/cpu/CPUJvmBuffer.kt
