+++ buffer/cpu/src/jvmAndroidMain/kotlin/com/wsr/cpu/elementwise/compare/JCompare.kt
