+++ buffer/cpu/src/jvmAndroidMain/kotlin/com/wsr/cpu/elementwise/math/JMath.kt
