+++ buffer/cpu/src/jvmAndroidMain/kotlin/com/wsr/cpu/elementwise/operation/times/JTimes.kt
