+++ buffer/cpu/src/jvmAndroidMain/kotlin/com/wsr/cpu/index/JIndex.kt
