+++ buffer/cpu/src/jvmAndroidMain/kotlin/com/wsr/cpu/reduction/JCollection.kt
