+++ buffer/cpu/src/jvmAndroidMain/kotlin/com/wsr/cpu/shape/JShape.kt
